extra/Veci, test improvements - Flesh out VL[Vec]234i in extra/ - Fixes so i/f headers can be combined - Add extra/234i test setup - Add Print234 support to extra/ headers - Add no_io target that excludes stdio/iostream from generated files - Remove inv() for int types to avoid accidental use - int vararg constructors - VLTestInt in test

+++ extra/VLVec234i.cpp
@@ -1,11 +1,11 @@
 //
-// VLVec234f.cpp
+// VLVec234i.cpp
 //
 // Andrew Willmott
 //
 
 
-#include "VLVec234f.hpp"
+#include "VLVec234i.hpp"
 
 
 
@@ -15,7 +15,7 @@
 
 
 
-Vec4f& Vec4f::MakeUnit(int n, float k)
+Vec4i& Vec4i::MakeUnit(int n, int k)
 {
     if (n == 0)
     { x = k; y = vl_zero; z = vl_zero; w = vl_zero; }
@@ -31,41 +31,41 @@ Vec4f& Vec4f::MakeUnit(int n, float k)
     return *this;
 }
 
-bool Vec4f::operator == (const Vec4f& a) const
+bool Vec4i::operator == (const Vec4i& a) const
 {
     return x == a.x && y == a.y && z == a.z && w == a.w;
 }
 
-bool Vec4f::operator != (const Vec4f& a) const
+bool Vec4i::operator != (const Vec4i& a) const
 {
     return x != a.x || y != a.y || z != a.z || w != a.w;
 }
 
-bool Vec4f::operator < (const Vec4f& a) const
+bool Vec4i::operator < (const Vec4i& a) const
 {
     return x < a.x && y < a.y && z < a.z && w < a.w;
 }
 
-bool Vec4f::operator >= (const Vec4f& a) const
+bool Vec4i::operator >= (const Vec4i& a) const
 {
     return x >= a.x && y >= a.y && z >= a.z && w >= a.w;
 }
 
-Vec4f cross(const Vec4f& a, const Vec4f& b, const Vec4f& c)
+Vec4i cross(const Vec4i& a, const Vec4i& b, const Vec4i& c)
 {
-    Vec4f result;
+    Vec4i result;
 
-    result.x =  dot(Vec3f(a.y, b.y, c.y), cross(Vec3f(a.z, b.z, c.z), Vec3f(a.w, b.w, c.w)));
-    result.y = -dot(Vec3f(a.x, b.x, c.x), cross(Vec3f(a.z, b.z, c.z), Vec3f(a.w, b.w, c.w)));
-    result.z =  dot(Vec3f(a.x, b.x, c.x), cross(Vec3f(a.y, b.y, c.y), Vec3f(a.w, b.w, c.w)));
-    result.w = -dot(Vec3f(a.x, b.x, c.x), cross(Vec3f(a.y, b.y, c.y), Vec3f(a.z, b.z, c.z)));
+    result.x =  dot(Vec3i(a.y, b.y, c.y), cross(Vec3i(a.z, b.z, c.z), Vec3i(a.w, b.w, c.w)));
+    result.y = -dot(Vec3i(a.x, b.x, c.x), cross(Vec3i(a.z, b.z, c.z), Vec3i(a.w, b.w, c.w)));
+    result.z =  dot(Vec3i(a.x, b.x, c.x), cross(Vec3i(a.y, b.y, c.y), Vec3i(a.w, b.w, c.w)));
+    result.w = -dot(Vec3i(a.x, b.x, c.x), cross(Vec3i(a.y, b.y, c.y), Vec3i(a.z, b.z, c.z)));
 
     return result;
 }
 
-Vec3f proj(const Vec4f& v)
+Vec3i proj(const Vec4i& v)
 {
-    Vec3f result;
+    Vec3i result;
 
     VL_ASSERT_MSG(v.w != 0, "(Vec4/proj) last elt. is zero");
